Update hex_to_ascii.py
I'm not entirely sure why you had some code that you did. I cleaned up what made sense to me, and made the code account for, and handle newlines. This is particularly a problem if you are working with hex values from a hexdump, or if sublime is convinced that the file you are working on is a binary one. Essentially, the values were presented as such:
xxxx xxxx xxxx xxxx xxxx xxxx xxxx xxxx
xxxx xxxx xxxx xxxx xxxx xxxx xxxx xxxx
...

This would throw an error, but the this proposed change will handle it.